+++ hadoop-hdds/framework/src/main/java/org/apache/hadoop/hdds/utils/db/RDBStoreCodecBufferIterator.java
@@ -115,7 +115,8 @@ class RDBStoreCodecBufferIterator
 
   @Override
   Table.KeyValue<CodecBuffer, CodecBuffer> getKeyValue() {
-    return RawKeyValue.create(key(), valueBuffer.getFromDb());
+    assertOpen();
+    return Table.newKeyValue(key(), valueBuffer.getFromDb());
   }
 
   @Override
